fix(cozy-intent): Use service as ref, not props
this will avoid race conditions where child components try to
use the context value (service) while it has its initial `undefined`
value. Here, it will always be defined, and the methods will be updated
as soon as the render cycle allows. Technically it should not
be possible for a race condition to occur where a child
tries to call a method that isn't injected yet, because the
serviceRef is instantiated with the first localMethods provided
as soon as the component is created