fix(harvest): Check the flag during React Lifecycle
It seems that sometime we can display the `DumbLaunchTriggerCard`
component even if the flag is on. The result is that the user
doesn't have, in that case, anything to manage the account because
elsewhere the flag is read during the React lifecycle.

I think this is why we can't reproduce it everytime, it's a
race condition between the loading of the flags and the read we do
here.

In order to read the flag when the flag's store is populated, we
need to read if within the React Lifecycle.

It should fix the issue.