feat(cozy-flags): Output types in dist
This commit enables automatic type generation.

In order to allow the automatic generation some blocking errors
had to be solved, which means `connect.js` had to be fully
converted to TypeScript.
MicroEE also needed to be typed in order to use the flag.store with
the now typed `connect.tsx`.
No runtime modification is expected or wanted during this upgrade.
This modification allows us to delete d.ts files existing in other
cozy repos that were typing cozy-flags themselves.
Removing bloat on other repos and adding types on source libs
should be overall a net gain for the quality of the codebase