feat: Add context to store Intent data
Like that, Harvest knows if it was called from an intent or not. And can
use the data to do what it wants.

I don't think this is a good idea to have this knowledge or having to
check if Harvest is called inside an intent or not, we should not have
to check that at this level.

But ATM, I don't know yet how to handle the opening of Drive when called
from an intent. If we just use a link then this is the intent that
changes the URL, not the caller. And just changing this intent url, then
we just close the intent and thats it.

One solution can have to call intent.redirect() for instance, but we need
to check if we are in an intent or not. So I don't like it neither.

For now, I just target blank because it fixes my issue.

BREAKING CHANGE: If you import Harvest's component without using the
HarvestWrapper or the Route Component, you have to wrap manually the
component in the IntentProvider.