fix: No more disabled submit button on 2nd 2FA
When the konnector was asking 2FA code multiple times, the user could
not submit the 2nd code since the submit button was forefever disabled.

This bug was due to the dependency on "konnectorRunning" in the flow
state. I don't know why this dependency was needed and neither the git
history or unit tests would help me.

But with this condition removed, every 2FA code are asked properly.
I also changed the related unit test to have more tests on this case
(which failed silently before).