feat: Add paper version detection
This introduces the possibility to detect a paper version,
based on the 2 sides OCR + reference rules.
Those rules can be applied on front and/or back side.
All the rules must be respected in order to detect a version.
For now, the rules are simple regex, evaluated against the
OCR.