fix: Don't display the KonnectorUpdateButton if maintenance
Since we can't launch a konnector is maintenance, there is no
need to try to update the konnector at this time.

This update button is still needed for manual upgrade
(permission changes for instance). Maybe we should only display
it in that case because for other case, the konnector will be
updated automatically before the launch.