feat: Display Scan result page when upload file by flagship
When uploading (via flagship), we no longer want to
arrive after the first Scan step, but rather display the
result of this first step, in order to be able to rotate the
image if necessary.

Move `getLastFormDataFile` in a useEffect because
during a flagship upload, the first rendering does not
have the file in the formData.