fix: Minimize services when cozy-notifications is used
cozy-notifications uses mjml to generate mail templates. mjml can not
be minimized as is. So previously, we disabled minimize for the whole
services configuration. It could lead to huge services bundle but it
was working.

We discovered that configuring Terser plugin with keep_fnames to true
is enough to allow minimization without breaking mjml. So let's use
it to reduce services bundle size.

From the Terser plugin documentation :
keep_fnames (default false) -- Pass true to not mangle function
names. Pass a regular expression to only keep function names matching
that regex. Useful for code relying on Function.prototype.name. See
also: the keep_fnames compress option.