fix(sharing): Handle memory leak in self removal
- A memory leak / bad component rerender in
sharingProvider seems to lose react props when
removing self from a sharing.
As a result, this will cascade in the react tree and end
in an unhandled crash that will kill the app completely (blank page).
- To fully handle this issue there is also something to handle
on cozy drive side
- Still, the root cause of the error has not been found yet.
But the error is severe enough to need this local fix