fix: Use sourceAccountIdentifier and konnector slug to show viewer modal
Parameter of useDataCardFiles change but we didn't update ViewerModal properly. I chose to search this param with a query to io.cozy.accounts as we had already this information without changing route path.  This is a temporary fix and I think we need to refactor the viewer part to have an abstraction to account in this modal