fix(mespapiers): Save the right information in contractType metadata
This commit fixes the following 2 cases:

- When you select the "Other" choice, then enter
a custom value and finally click again on "Other
 (or select another option then return to " Other"),
the saved value is `Other` instead of the custom value.

- When you select the "Other" choice, then enter
a custom value, then delete it, the saved value is `""`
in place of the `Other` value.

Also, the `useEffect` only useful for setting the default value,
needed in this current implementation,
does not need to react to dependency value change.
This is basically bad practice, but I don't see how to do otherwise
without reviewing the implementation more fully.