fix: Do not save identifiers in vault when no vendor_link
The konnector identifiers won't be saved in vault when
the konnector's manifest does not have any vendor_link.

This avoids to have the whole list of identifiers displayed to
the user when he tries to configure a konnector without
any vendor_link, like NextCloud.

BREAKING CHANGE: you must have `cozy-ui >= 103.12.0`