fix: Transmit correct account identifiers on reconnect
When the user tries to connecteur a server konnector with
wrong identifiers and fix identifiers and retry conection,
the  previous identifier a resent to the konnector and an
io.cozy.account document, not linked to any trigger is also
created for each atempt.

Now the trigger manager will reuse the account from the
the current ConnectionFlow instance when possible.