feat: Load new data when webview killed in 'scanDocument'
The 4rd step when the flagship app reload the Mes Papiers app after
the mobile operating system killed the webview during a
'scanDocument' webview intent is to load the new data to be at the
same point as if the webview was never killed. It is possible because
only the webview is killed, not the main process of the flagship
which is doing the 'scanDocument' webview intent.

fix: Webview killed