fix: Wait until StepperDialogProvider is ready
There are multiple independant initialization processes in the
StepperDialogProvider :
- set currentDefinition
- set allCurrentSteps
- set currentStepIndex

It can lead to strange behavior in the FormDataProvider which depends
on data from the StepperDialogProvider. So let's wait that the
StepperDialogProvider is correctly initialized. It is a very simple
way to avoid strange behavior and unnecessary rerender and it does
not have a huge cost because the initialization processes are fast.