test: Remove StepperDialogProvider from AppLike
We now need to mock some helper calls in
StepperDialogProvider to pass the "isReady" state.
StepperDialogProvider was added in AppLike, so we needed to mock
these calls for *every* test.

To avoid this, I remove StepperDialogProvider from AppLike and add it
in every component that use it. It is more representative of what
happen in real life because Stepper DialogProvider is only present
for the CreatePaperModal.