feat: Download file on mobile viewer on press
We want to download file on mobile viewer on press (so when we put
the finger on the screen) and not on tap (so when we put the finger
on the screen and we release it).

I added a throttle to avoid download multiple time the same file
because we can no use anymore 'tapCount' with press which was used to
avoid download 3 times the file if we tapped 3 times.

I also moved to useMemo instead of useCallback to support throttle
(see https://github.com/facebook/react/issues/19240).