feat: Use local search only for peristent sessions
We now rely on the checkbox on login to determine whether or not the
local search should be used. If the user enable this checkbox, it means
he trusts this device and accept to store local data on it.
Otherwise, we do not store any private data, and directly query the
stack to build the search index.

Note the update mechanism is the same and rely on realtime. The
difference is that we do not replicate the document for non local
search.