feat: Do not use flexsearch store
Flexsearch has a built-in store capability, in order to store the full
content of indexed documents.
This is useful, however, this takes extra memory for nothing, as we
already have the cozy-client store. All the indexed documents for the
search are supposed to exist in cozy-client store, as we query them from
Pouch at init, and retrieve updates from realtime, already plugged in
cozy-client.