fix: Revert "feat: Do not use flexsearch store"
This reverts commit 104ea850ef9b2bf5ae5094f5d10f81bf8872bcf1.
As we dynamically compute the path, that is not stored on PouchDB, it
is required to store it in the cozy-client store and thus update all
the files accordingly with the path.
However, this is not usual to do so and might lead to unexpected
side-effects.
So for safety, we revert this commit for now and might rethink this
later.