feat(Sharing): Add isPublic prop to SharingProvider
In public mode, execution of the promise with calls
to the following methods:
`this.sharingCol.findByDoctype()`,
`this.permissionCol.findLinksByDoctype()` and
`fetchApps()`
failed (cf 403).

So we have useless calls and
the `fetchAllSharings` method which in the end does nothing.

The `isPublic` prop quickly exits `fetchAllSharings`
to avoid 3 useless requests.

Real time is retained as functional and useful.
(Links sharing and Cozy to Cozy added but not sync)