feat: Add NativeFileSharingProvider and shareNative action
Add a provider and a action to expose an ActionMenu action that can
share files with mobile device share capabilities.

shareFilesNative from the NativeFileSharingProvider must be passed to
the action.

Inspired from code in mespapiers. It will be used in cozy-drive,
cozy-notes and mespapiers.

BREAKING CHANGE: you now need cozy-intent >=2.29.1
and cozy-device-helper >= 3.7.1 with cozy-sharing