fix: Imprecise search results
Sometimes a search could return surprising results, with no clear match.
For instance a "test" search would return a file with the name
"dist.pdf".
This is because we used the "balance" encoder, which performs a
"literal" transformation on text, e.g. "d" matches "t", "i" matches "e",
and so on. Thus, "dist" is transformed into "test", hence the matching.

This is not wanted in our case, and makes the index bigger, so we
disable it.