feat(sharing): Improve BoxPassword
As we can't display the previously saved password,
we'd like to display a specific placeholder
that improves the UX.

We'll also make sure that the password value is `“”`
and not `undefined` when we deactivate it.
This makes it clear to the stack that we want to delete
the password and not keep it unchanged.