fix: Improve indexing time
When indexing several docs, we used to call the `computeFileFullpath`,
which is useful to compute the path of all files, by querying the
parent directory.
However, this async method is not useful after we queried all the
required docs, as we already have the parent docs.
So, we used instead the `addFilePaths` method.
Thanks to it, the indexing should be speed up, and we avoid a
useless async method.
Note the `void this.indexDoc` call was also making unreliable
measure times, as it was actually impacting the following time
measures.