feat: Import svg as React components with rsbuild
With webpack, we relied on svg-sprite-loader to import SVGs. It
created an element in the DOM for each SVGs and cozy-ui Icon was
built to understand and display these SVGs by creating a xlink:href
to these SVGs in the DOM.

svg-sprite-loader does not work with rsbuild. Instead, by default,
rsbuild can :
- inline SVG
- import SVG as static asset

But in both case, we can not easily change the color of the SVG which
is necessary in cozy-ui Icon.

By using the SVGr rsbuild plugin, rsbuild can import SVG as React
component which are understandable by cozy-ui Icon and we can change
the color. It works as is with cozy-ui Icon.