feat: Do not rebuild search index from scratch
Each time a searchEngine is instanciated, we used to rebuild the search
indexes from scratch, i.e. query the local database and compute the
indexes.  This can be quite time consuming, especially in slow mobile
devices. Therefore, we now export the search indexes on disk, and import
them at init time.  We save the lastSeq for each index, to be the
starting point of the next changes batch updates, fetched from the local
database after a remote replication.

Some measures, performed on a 8K files database:

**Redmi Note 10**

- Create index: without peristance - old way
  - Query all io.cozy.files : 3700ms
  - Build files index : 6500ms
  - **Total: 10200 ms**
- Create index: with perstance - new way
  - **Import persisted index: 800 ms**
  - Export index: 1057ms

**OnePlus Nord 4**

- Create index: without peristance - old way
  - Query all io.cozy.files : 2000ms
  - Build files index : 2900ms
  - **Total: 4900 ms**
- Create index: with perstance - new way
  - **Import persisted index: 332 ms**
  - Export index: 363ms

If we compare the query+build index time w.r.t the import time, we have
a ~13x time improvement before the index is ready to use.