feat: Correctly deal with paths
Now that the docs are not stored inside index, we need to compute and
maintain file paths.  We have 3 possibilities here:
1. Store paths in CouchDB
  - Can be costful
  - Looks risky because it will produce new revisions and might end up
    in replication conflicts
2. Store paths in store
  - Require new API to allow this in cozy-client
  - Differences between db and store that might lead to overwriting
3. Maintain a dedicated structure in Search
  - Less generic
  - Easier to implement

We go for 3. as the need is only for search for now. But we might move
it to cozy-client if we decide to extend the local queries more broadly.