feat: Use parent origin instead of document.referrer
We sent message from iframe to origin infered document.referrer
because document.referrer, when in an iframe, is set initially to the
iframe URL. But if a redirection happens in the iframe (which happens
during an oauth login) document.referrer is updated. So we sent
message from iframe to the wrong origin.

Here we introduce a way to ask the parent window, if it exists, his
origin. We just take the origin from the parent window answer
metadata. We do not trust the parent window.

By doing this, bridge will be able to work even after a reload or a
redirection inside the iframe.