feat: Correctly init DataProxy
We now render the `DataProxyProvider` children only when the DataProxy
is  correctly initialized on the web.
To know this, we rely on the cozy-web-dataproxy app to send a "ready"
message once it is indeed ready.
This is meant to avoid some timing issues, where the app is ready to
request the DataProxy, but the DataProxy is not initialized yet.