feat(cozy-external-bridge): Add a bridge route prefix
Now, when synchronizing URL, we will add a prefix in the parent app.
So for example, if iframe go from iframe.com to iframe.com/welcome we
will update parent URL to parent.com/bridge/welcome instead of
parent.com/welcome. It will allow parent app to have its own routing
(for example for cozy-bar AI assistant).

It simplifies useInitialRedirection code because now we can easily
know if parent URL need to be forwarded to iframe or not.