fix: Redirect on load display blank page
iframe.src can be null since
https://github.com/cozy/cozy-twakechat/pull/20 so useRedirectOnLoad
was throwing "Invalid URL" error when constructing the destination
url with URL constructor and iframe.src.

Here we change the way of doing this and useExternalBridge does not
redirect the iframe directly anymore (which was a bad idea) but
instead build the URL and just return it. We do not build the URL
from the iframe.src but from the origin received in parameters which
are the same.

Some tests are renamed because I forgot to rename them previously.

BREAKING CHANGE: useExternalBridge hook does not redirect
the iframe automatically if necessaru and returns instead an
`urlToLoad` url in its return value