fix: Query local PouchDB queries rather than store
We used to query the store on the whole doctype to get docs by id, at
search response time.  We noticed in the past that is was much more
efficient that querying PouchDB directly. However, after many
improvements in cozy-pouch-link, this appears as not the case anymore:
querying 100 docs by id takes few ms.

Furthermore, we now prefer to disable the store for dataproxy queries,
to both preserve RAM and keep good performances across time, as
using the store might hurt query efficiency eventually, as store
evaluation can be costful and grow over time. Especially since we want
to rely on the DataProxy to evaluate cozy-client queries from the app
side.
As a side effect, this was breaking web search because we tried to
query the store while it was disabled by cozy-web-dataproxy

See https://github.com/cozy/cozy-web-data-proxy/commit/
2616304af186e95b9e11db8174debd0a2de73e45