feat: Use specific extension for Docs notes
Using `.md` extension for Docs notes makes sense, as it is indeed a
markdown that is exported as file content.
However, this can be problematic in case the user uses the desktop
client and ends up editing the markdown in an external editor: the
markdown will be overwritten by Docs the next time the note is opened in
the app.

To prevent such situation, we use a special `.docs-note` in the same
spirit than `.cozy-note`. However, we keep the `text/markdown` mimetype
to identify the file as true markdown.