feat: Improve sharings retrieval performances
When dealing with large sharings, requesting the list of sharings can be
quite slow, i.e. several seconds.  This is because by default this route
returns the list of *all* shared docs. So if we deal with thousands of
docs, this is slow to compute and transmit.

We do not need this list of shared docs, so let's disable it.