feat: Handle 2 characters edge cases
With the current search setting, this case was not working:

```
filename: "UI-UX Guideline.cozy-note"
search: "UI-UX"
```

This is because the tokenizer takes the `-` as a separator and we end up
with ["UI", "UX"] tokens. As the minimal word length was 3, they were
skipped from the indexation.

So, we now specify a 2 minimal word length to handle this kind of cases.
Note this will slightly increase the index size