fix(cozy-dataproxy-lib): Avoid blank screen when dataproxy not available
Since version 4.2.0, we wait dataproxy services to be ready to render
children to be sure that we don't miss call to dataproxy methods.

But if dataproxy services are not available at all (so if
flag('cozy.search.enabled') was not set or if dataproxy app was
missing), we were never rendering the children leading to a blank
screen.

So now, we render children if there is no dataproxy OR when dataproxy
is ready.