feat: Expose reconnect and stop methods in plugin
Clients dealing with realtime might need to have control over websocket
connection.  This is typically useful for the dataproxy shared worker,
to reconnect realtime after connectivity loss.  This is normally handled
by cozy-realtime directly, through the `online` event:
https://github.com/cozy/cozy-libs/blob/e65ef0b285982ef58269f766870bcb26fee91ef4/packages/cozy-realtime/src/CozyRealtime.js#L57

But this event is not available in the shared worker context.  Thus, we
want to have the possibility to manualy reconnect from there.