feat: Replicate data at init
Useful to have fresh data. We used to avoid this replication at init
because it can be a costful operation, and we didn't want to slow down
the mobile app opening.
But, this is already avoided through the `shouldInit` option, which is
currently disabled for the mobile app, and enabled for the web.