fix: Update file paths after move or rename
When a file is renamed or moved, the fullpath changes.  When the path is
not stored, typically with PouchDB, it is retrieved from an in-memory
Map to be efficient. Hence, this path must be updated in this Map, which
was not done, causing wrong path display after a move/rename.