feat: Always render dataproxy children
Previously, the rendering of `DataProxyProvider` children were
conditioned by the initialization of the dataproxy.

It was because we wanted to avoid timing issues, were the App started to
use the DataProxy before it was actually ready.  See
https://github.com/cozy/cozy-libs/commit/
bdb16705103fa222d631750403d951eee61f4868

We later made it less strict, because if anything make the dataproxy not
initialized, nothing is displayed on the app side with no logs, making
it hard to debug for the developer.  See
https://github.com/cozy/cozy-libs/pull/2801

However, we still met situations with no rendering, because the
dataproxy was not available at all:

- Public sharing: https://github.com/cozy/cozy-drive/pull/3439
- Unsupported shared workers:
  https://github.com/cozy/cozy-web-data-proxy/pull/36

So now, we always render children, and expect the `DataProxyLink` to
handle queries made on non-ready DataProxy by queuing them:
https://github.com/cozy/cozy-client/pull/1622