feat: Display warning for long commit body lines
We used to have an error for commit body lines > 72. This is quite
annoying because it forces the dev to redo the commit. Furthermore, this
rule is not always legitimate, typically when one when to put a long
URL, forcing to add a line break in the URL.

So let's display a warning rather than an error.