feat(sharing): Add shared drive edit modal
This modal is opened when we open the usual sharing modal in a shared
drive. So it should work as is directly in drive.

For the moment, shared drive modal is a lot different than cozy to
cozy sharing modal that's why I decided to create a new one inside
ShareModal.