fix: Infinite render when default option
The app mouting the Provider can provide options. If no option is
provided, a default `{}` is given.
The reference of this `{}` will change everytime, so if we pass it in
dependency array on a `useEffect`, it will be trigger everytime, making
an infinite loop.
This typically happened with Home, that does give any option, but not in
Drive, that provide options.